Refactor Footer Components
