+++ apps/filecoin-site/src/app/_components/Footer/FilecoinFoundationLogo.tsx
